+++ src/lib.rs
@@ -47,6 +47,7 @@ doc_comment::doctest!("../README.md");
 
 pub use scip_sys as ffi;
 pub mod branchrule;
+pub mod pricer;
 pub mod constraint;
 pub mod model;
 pub mod retcode;