+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Solving stats methods for number of nodes, time, and number of lp iterations. 
 - Branching rule plugin. 
+- Variable pricer plugin. 
 ### Fixed
 ### Changed
 ### Removed